Virtual Machine (#22)
* Added support for acronyms

This change adds code generation support for acronyms (e.g. VM -> vm instead of vM).

* Added client reference to response objects

This change is done so that later functions that perform API calls can be implemented directly on the response objects. This makes it easier to call them.

* Regenerated code with go generate

* Minor fixups

Added minor changes that didn't fit anywhere else.

* Added datacenter

The datacenter is the main object linking clusters and networks, so we need it for VM creation.

* Added networks

Networks are required for later NIC support.

* Added VM creation

Added basic, empty VM creation. This will be extended in the future with additional parameters.

* Added VNIC profiles

VNIC profiles are templates for network interfaces.

* Added NICs

NICs represent virtual network interfaces on VMs. It is one of many objects that will be added to VMs.

* gofmt

Ran gofmt.

* Updated readme

* Fixed TLS bug

Fixed a TLS bug where patterns wouldn't be stored in pointers.

* Added better test helper

This test helper automatically identifies test parameters from environment variables.

* Added optional VM creation parameters

This adds an parameter to include optional VM parameters on creation.

* Documentation update

Adding a lot of API doc making it easier for users to navigate the client.