Stricter linting (#23)
* Stricter linting

This change enables and documents a number of linters with golangci-lint.

* Added unparam rule.

This change adds the unparam rule, requiring unused parameters to be marked explicitly with _ to prevent bugs.

* Added godot rule.

Added godot rule that requires doc comments to end in a dot. This helps prevent incomplete comment lines, such as half-written sentences.

* Added gochecknoglobals rule.

This rule prevents accidentally adding global variables and introducing global scope in an object-oriented environment.

* Added gochecknoinits rule.

This rule prevents init() functions to be added.

* Added testpackage rule.

This rule enforces tests to be put in the _test package, except where necessary.

* Added revive rule.

Added the revive rule, which is a replacement for the defunct golint.

* Added nestif rule.

This rule complains about deeply nested if cases, resulting in unreadable code.

* Added nakedret rule.

This rule prevents naked returns on long functions for better readability.

* Added gocritic rule.

Gocritic is a tool that complains about constructs that could be written simpler. This makes the code more readable.

* Added gosec rule.

This rule detects potentially security-relevant misconfiguration in Go code.

* Added gofumpt and whitespace rules.

Gofumpt and whitespace together create a more strict code formatting ruleset that can be automatically applied.

* Final cleanup in configuration file

* Fixed ifshort error