Fixes #110: No more failure on missing initialization
Before this fix the GET / LIST calls for VMs would fail with the message "no initialization field found on VM object" if there is no initialization set. However, creation would not fail. This change fixes that issue and adds a test to make sure this is actually fixed. The test has been verified against the oVirt engine.